add functions measuring single channel (ADC) and  function of 1byte data in HK (HKDownlinkFR2), change changeCharMorse function type from int to long (succeed in build but not check with board)

+++ PIC_Tx/FMCW.h
@@ -40,7 +40,7 @@ void HKDownlinkFR2(void);
 
 /*Function : CW downlink*/
 void commandSwitchCWDownlink(UBYTE, UBYTE, UBYTE, UBYTE, UBYTE, UINT, UBYTE);
-int changeCharMorse (char);
+long changeCharMorse (char);
 char changeBinaryToChar(UBYTE);
 void DevideDataAndChangeBinaryToChar (UBYTE, UBYTE*);
 void sendMorse(char*,size_t);
